move CheckpointSerializer and make it public
